Add an context-dependent allocator interface and two allocators
This makes it possible to add additional allocators over the
main system allocator, and then select the appropriate allocator
depending on the context of the caller.   Currently, the context
is the calling thread.  If the allocator is called in interrupt
context, the system allocator is used, even if the thread has
a custom allocator.

See include/nautilus/alloc.h for the implementation and user
interface