Fix FPU-on-exception-handling
Somehow when this got transferred from the other branch, the exception
handling code ended up saving GPRs *after* the FPU state save.   This
is bad because the FPU state save uses the GPRs, so the thread GPRs
get mangled.   This is a problem for, say, a page fault, because you
will come back from a page fault.   With your registers ever so slightly
mangled.